+++ src/main/java/application/service/user/UserService.java
@@ -5,7 +5,7 @@ import application.rest.domain.UserDTO;
 import application.service.DatabaseServiceInterface;
 import application.service.response.ServiceResponse;
 
-public interface UserService extends DatabaseServiceInterface<User, Integer, UserDTO> {
+public interface UserService extends DatabaseServiceInterface<User, Long, UserDTO> {
 
     ServiceResponse<UserDTO> read(String username);
