+++ src/main/java/application/service/user/UserServiceImpl.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class UserServiceImpl extends AbstractDatabaseService<User, Integer, UserRepository, UserDTO> implements UserService {
+public class UserServiceImpl extends AbstractDatabaseService<User, Long, UserRepository, UserDTO> implements UserService {
 
     @Autowired
     private UserRepository userRepository;
